cleanup: rename client_options* files.
The main class in them is ConnectionOptions, so the file should match.
This change also moves the other classes in that header to (a new file)
called `read_options.h`.
