+++ google/cloud/spanner/connection_options.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CLIENT_OPTIONS_H_
-#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CLIENT_OPTIONS_H_
+#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CONNECTION_OPTIONS_H_
+#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CONNECTION_OPTIONS_H_
 
 #include "google/cloud/spanner/version.h"
 #include "google/cloud/status_or.h"
@@ -83,36 +83,9 @@ class ConnectionOptions {
   std::set<std::string> tracing_components_;
 };
 
-/// Options passed to `Client::Read` or `Client::PartitionRead`.
-struct ReadOptions {
-  /**
-   * If non-empty, the name of an index on a database table. This index is used
-   * instead of the table primary key when interpreting the `KeySet`and sorting
-   * result rows.
-   */
-  std::string index_name;
-
-  /**
-   * Limit on the number of rows to yield, or 0 for no limit.
-   * A limit cannot be specified when calling`PartitionRead`.
-   */
-  std::int64_t limit = 0;
-};
-
-inline bool operator==(ReadOptions const& lhs, ReadOptions const& rhs) {
-  return lhs.limit == rhs.limit && lhs.index_name == rhs.index_name;
-}
-
-inline bool operator!=(ReadOptions const& lhs, ReadOptions const& rhs) {
-  return !(lhs == rhs);
-}
-
-/// Options passed to `Client::PartitionRead` or `Client::PartitionQuery`
-using PartitionOptions = google::spanner::v1::PartitionOptions;
-
 }  // namespace SPANNER_CLIENT_NS
 }  // namespace spanner
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CLIENT_OPTIONS_H_
+#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_CONNECTION_OPTIONS_H_
